refactor: reorganizar importaciones en el hook de checkout
