+++ src/components/checkout/CheckoutFormSimple.tsx
@@ -10,7 +10,8 @@ import { Checkbox } from '@/components/ui/checkbox';
 import { Separator } from '@/components/ui/separator';
 import { Badge } from '@/components/ui/badge';
 import { CreditCard, MapPin, Lock, CheckCircle } from 'lucide-react';
-import { CartItem, User as UserType } from '@/types';
+import { User as UserType } from '@/types';
+import { CartItem } from '@/context/CartContext';
 import { useCheckout } from '@/hooks/useCheckout';
 import { checkCartStock } from '@/services/inventory.service';
 import { validateCardNumber } from '@/utils/validation';
